+++ src/syntax/config.js
@@ -1,0 +1,33 @@
+module.exports.body = function(theme) {
+  let ini = [
+    {
+      name: "[INI] - Entity",
+      scope: "source.ini entity",
+      settings: {
+        foreground: theme.default + "cc"
+      }
+    },
+    {
+      name: "[INI] - Keyword",
+      scope: "source.ini keyword",
+      settings: {
+        foreground: theme.yellow
+      }
+    },
+    {
+      name: "[INI] - Punctuation Definition",
+      scope: "source.ini punctuation.definition",
+      settings: {
+        foreground: theme.yellow
+      }
+    },
+    {
+      name: "[INI] - Punctuation Separator",
+      scope: "source.ini punctuation.separator",
+      settings: {
+        foreground: theme.yellow
+      }
+    }
+  ];
+  return Object.assign([].concat(ini));
+};