+++ CHANGELOG.md
@@ -4,16 +4,17 @@
 
 ### New features
 
+- Publish this changelog (never too late) 🎉
 - Add yml support
 - Add xml support
 - Add jsx support
 
 ### Enhancements
 
-- Publish this changelog (never too late) 🎉
 - Better CS support
 - Better Java support
 - Better Vue support
+- Better TypeScript support
 
 ### Fixes